Этап 2

+++ stat-service/stat-server/src/main/java/ru/practicum/statistic/api/storage/StatisticInfo.java
@@ -5,5 +5,5 @@ public interface StatisticInfo {
 
     String getUri();
 
-    long getHits();
+    int getHits();
 }
